% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql2json.R
\name{graphql2json}
\alias{graphql2json}
\title{Dump GraphQL to JSON}
\usage{
graphql2json(input)
}
\arguments{
\item{input}{a string with graphql syntax}
}
\description{
Parses GraphQL and exports the AST in JSON format.
}
\examples{
graphql2json("{ field(complex: { a: { b: [ $var ] } }) }")
}

