% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql2json.R
\name{schema2json}
\alias{graphql_schema}
\alias{schema2json}
\title{Dump All GraphQL to JSON}
\usage{
schema2json(input)
}
\arguments{
\item{input}{a string with graphql syntax}
}
\description{
Parses GraphQL queries and schema definitions and exports the AST in JSON format.
}
\examples{
graphql2json("schema { query: QueryType }")
}

