% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphql2json.R
\name{graphql2json}
\alias{graphql2json}
\alias{graphql}
\title{Dump GraphQL Queries to JSON}
\usage{
graphql2json(input, parse_schema = FALSE)
}
\arguments{
\item{input}{a string with graphql syntax}

\item{parse_schema}{boolean to enable schema defintion parsing}
}
\description{
Parses GraphQL queries and exports the AST in JSON format.
}
\examples{
graphql2json("{ field(complex: { a: { b: [ $var ] } }) }")
graphql2json("schema { query: QueryType }", TRUE)
}
